Added Flexible Polyline Decoder
Added minitests and using BigDecimal instead of float because float = bad
